feat: move types to single types file
