+++ components/general/select/index.ts
@@ -1,1 +1,2 @@
 export * from "./SingleSelect";
+export * from "./SelectTypes";
